requirements.txt
ollama
chromadb
sentence-transformers
streamlit
pymupdf
pysqlite3-binary